+++ router/user.js
@@ -15,6 +15,7 @@ router.post('/register', expressJoi(user_schema), login_handle.register)
 // 用户登录
 router.post('/login', expressJoi(user_schema), login_handle.login)
 
-
+// 用户页面相关模块
+router.get('/user/index',login_handle.handleUserIndex)
 // 导出
 module.exports = router